# apps/web/public/.well-known/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RootFetch Public API",
    "version": "1.0.0",
    "description": "Read-only, artifact-backed API surface for MCP discovery and immutable run artifacts."
  },
  "servers": [
    {
      "url": "https://rootfetch.com"
    }
  ],
  "paths": {
...
    "/api/mcp/stats": {
      "get": {
        "summary": "MCP usage stats",
        "operationId": "getMcpUsageStats",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 30,
              "default": 7
            },
            "description": "Window size in days"
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregated MCP usage stats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mode": {
                      "type": "string",
                      "enum": [
                        "shared",
                        "local"
                      ]
                    },
                    "generated_at_utc": {
                      "type": "string"
                    },
                    "window_days": {
                      "type": "integer"
                    },
                    "totals": {
                      "type": "object",
                      "properties": {
                        "requests": {
                          "type": "integer"
                        },
                        "rate_limited": {
                          "type": "integer"
                        },
                        "errors": {
                          "type": "integer"
                        }
                      }
                    },
                    "daily": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "date_utc": {
                            "type": "string"
                          },
                          "requests": {
                            "type": "integer"
                          },
                          "rate_limited": {
                            "type": "integer"
                          },
                          "errors": {
                            "type": "integer"
                          }
                        }
                      }
                    },
                    "adoption_kpi": {
                      "type": "object",
                      "properties": {
                        "unique_clients": {
                          "type": "integer"
                        },
                        "repeat_clients": {
                          "type": "integer"
                        },
                        "repeat_client_rate_pct": {
                          "type": "number"
                        },
                        "tool_call_requests": {
                          "type": "integer"
                        },
                        "tool_call_success_rate_pct": {
                          "type": "number"
                        },
                        "initialize_requests": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (admin auth required)"
          }
        },
        "security": [
          {
            "BasicAuth": []
          }
        ]
      }
    },
    "/api/mcp/events": {
      "get": {
        "summary": "MCP recent events",
        "operationId": "getMcpUsageEvents",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          },
          {
            "name": "rpc_method",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tool_name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recent MCP request events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "generated_at_utc": {
                      "type": "string"
                    },
                    "mode": {
                      "type": "string",
                      "enum": [
                        "shared",
                        "local"
                      ]
                    },
                    "events": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "ts_utc": {
                            "type": "string"
                          },
                          "http_method": {
                            "type": "string"
                          },
                          "rpc_method": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "tool_name": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "status": {
                            "type": "integer"
                          },
                          "duration_ms": {
                            "type": "integer"
                          },
                          "rate_limited": {
                            "type": "boolean"
                          },
                          "kind": {
                            "type": "string"
                          },
                          "ip_hash": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (admin auth required)"
          }
        },
        "security": [
          {
            "BasicAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "BasicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Admin telemetry endpoints require HTTP Basic Auth."
      }
    }
  }
}